+++ src/cron/task-run-history.ts
@@ -6,8 +6,13 @@ import {
 import { uniqueValues } from "@openclaw/normalization-core/string-normalization";
 import { listTaskRegistryRecordsByRuntimeSourceIdFromSqlite } from "../tasks/task-registry.store.sqlite.js";
 import type { TaskRecord } from "../tasks/task-registry.types.js";
-import { isCronDeliveryStatus, isCronRunStatus, type CronRunLogEntry } from "./run-log-types.js";
-import { cronTaskRecordStoreKey, cronTaskRecordToRunLogEntry } from "./task-run-detail.js";
+import type { CronRunLogEntry } from "./run-log-types.js";
+import {
+  cronTaskRecordStoreKey,
+  cronTaskRecordToRunLogEntry,
+  isCronDeliveryStatus,
+  isCronRunStatus,
+} from "./task-run-detail.js";
 import type { CronDeliveryStatus, CronRunStatus } from "./types.js";
 
 type CronRunHistorySortDir = "asc" | "desc";
